% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_weights.R
\name{kernel_weights}
\alias{kernel_weights}
\title{Gaussian kernel weights}
\usage{
kernel_weights(d, phi = 0.5, k)
}
\arguments{
\item{d}{data}

\item{phi}{some parameter, usually 0.5}

\item{k}{k newarest neighbours}
}
\description{
Gaussian kernel weighhts
}
\keyword{clustering}
\keyword{clustering,}
\keyword{graph}
\keyword{markov}
