% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovClust.R
\name{markovClust}
\alias{markovClust}
\title{Markov clustering function}
\usage{
markovClust(M, inflation = 2.5, expansion = 2, eStop = 0.001,
  max_it = 100)
}
\arguments{
\item{M}{adjacency matrix}

\item{inflation}{inflation parameter}

\item{expansion}{expansion parameter}

\item{eStop}{stop criterium - Frobenius norm of matrix M_t - M_{t-1}}

\item{max_it}{maximum number of iterations}
}
\description{
Markov clustering function algorithm
}
\examples{
A <- rbind(c(1, 1, 0, 1, 0, 1, 0),
           c(1, 1, 1, 1, 0, 0, 0),
           c(0, 1, 1, 1, 0, 0, 1),
           c(1, 1, 1, 1, 1, 0, 0),
           c(0, 0, 0, 1, 1, 1, 1),
           c(1, 0, 0, 0, 1, 1, 1),
           c(0, 0, 1, 0, 1, 1, 1))
D <- diag(rowSums(A))
M0 <- solve(D) \%*\% A
markovClust(M0, 2.5, 0.001)
}
\keyword{clustering}
\keyword{clustering,}
\keyword{graph}
\keyword{markov}
