% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MarkovClustObject.R
\name{summary.MarkovClustObject}
\alias{summary.MarkovClustObject}
\title{Summarizing Markov clustering algorithm}
\usage{
\method{summary}{MarkovClustObject}(MCO)
}
\description{
Summarizing Markov clustering algorithm
}
